+++ lib/src/chart/bar/container.dart
@@ -2,6 +2,8 @@ import 'dart:ui' as ui show Size;
 
 import 'package:flutter_charts/src/chart/options.dart';
 import 'package:flutter_charts/src/chart/data.dart';
+import 'package:flutter_charts/src/chart/iterative_layout_strategy.dart'
+  as strategy show LabelLayoutStrategy;
 import '../container.dart';
 
 import 'presenter.dart';
@@ -17,11 +19,13 @@ class VerticalBarChartContainer extends ChartContainer {
     ui.Size chartArea,
     ChartData chartData,
     ChartOptions chartOptions,
+    strategy.LabelLayoutStrategy xContainerLabelLayoutStrategy,
   })
       : super(
     chartArea: chartArea,
     chartData: chartData,
     chartOptions: chartOptions,
+    xContainerLabelLayoutStrategy: xContainerLabelLayoutStrategy,
   ) {
     presenterCreator = new VerticalBarLeafCreator();
   }