+++ lib/src/util/range.dart
@@ -3,9 +3,9 @@ import 'package:decimal/decimal.dart' as decimal;
 import 'package:flutter_charts/src/chart/options.dart';
 import 'util.dart' as util;
 
-// todo -11 - this library (range.dart) has been modified for Dart 2.0
+// todo -2 - this library (range.dart) has been modified for Dart 2.0
 //            using a hack which replaces all List<num> to List<double>,
-//            also some int replaced with double
+//            also some int replaced with double. Parametrize with T
 
 
 /// Scalable range, supporting creation of scaled x and y axis labels.
@@ -19,7 +19,8 @@ class Range {
 
   ChartOptions _options;
 
-  /// The auto-label generator [makeLabelsFromData] can decrease this but not increase. todo 00 - move to options
+  // todo -1 - move to options
+  /// The auto-label generator [makeLabelsFromData] can decrease this but not increase.
   int _maxLabels;
 
   /// Constructs a scalable range from a list of passed [values].
@@ -41,8 +42,7 @@ class Range {
 
   /// superior and inferior closure - min and max of values
   Interval get _closure => new Interval(
-      _values.reduce(math.min).toDouble(), _values.reduce(math.max).toDouble(), true, true);
-       // todo -11 _values.reduce(math.min), _values.reduce(math.max), true, true);
+      _values.reduce(math.min).toDouble(), _values.reduce(math.max).toDouble(), true, true); // todo -2 remove toDouble after parametrization
 
   /// Automatically generates unscaled labels (more precisely their values)
   /// from data.
@@ -183,65 +183,6 @@ class Range {
     return labels;
   }
 }
-
-
-/* // todo -11 repla
-// original version before replacing int to double where needed
-  List<double> _distributeLabelsIn(Interval interval) {
-    Poly polyMin = new Poly(from: interval.min);
-    Poly polyMax = new Poly(from: interval.max);
-
-    int powerMax = polyMax.maxPower;
-    int coeffMax = polyMax.coeffAtMaxPower;
-    int signMax = polyMax.signum;
-
-    // using Min makes sense if one or both (min, max) are negative
-    int powerMin = polyMin.maxPower;
-    int coeffMin = polyMin.coeffAtMaxPower;
-    int signMin = polyMin.signum;
-
-    List<double> labels = [];
-    int power = math.max(powerMin, powerMax);
-
-    // todo 1 refactor this and make generic
-    if (signMax <= 0 && signMin <= 0 || signMax >= 0 && signMin >= 0) {
-      // both negative or positive
-      if (signMax <= 0) {
-        for (int l = signMin * coeffMin; l <= 0; l++) {
-          labels.add(l * math.pow(10, power));
-        }
-      } else {
-        // signMax >= 0
-        for (int l = 0; l <= signMax * coeffMax; l++) {
-          labels.add(l * math.pow(10, power));
-        }
-      }
-    } else {
-      // min is negative, max is positive - need added logic
-      if (powerMax == powerMin) {
-        for (int l = signMin * coeffMin; l <= signMax * coeffMax; l++) {
-          labels.add(l * math.pow(10, power));
-        }
-      } else if (powerMax < powerMin) {
-        for (int l = signMin * coeffMin; l <= 1; l++) {
-          // just one over 0
-          labels.add(l * math.pow(10, power));
-        }
-      } else if (powerMax > powerMin) {
-        for (int l = signMin * 1; l <= signMax * coeffMax; l++) {
-          // just one under 0
-          labels.add(l * math.pow(10, power));
-        }
-      } else {
-        throw new Exception("Unexpected power: $powerMin, $powerMax ");
-      }
-    }
-
-    return labels;
-  }
-}
-
- */
 
 /// Encapsulating Y axis scaling (dataRange scaling to available pixels)
 /// and Y Labels creation and formatting.
@@ -390,7 +331,7 @@ class LabelInfo {
 
   /// Self-scale the RangeOutput to the scale of the available chart size.
   void _scaleLabelValue() {
-    // todo 00 consider what to do about the toDouble() - may want to ensure higher up
+    // todo -2 consider what to do about the toDouble() - may want to ensure higher up
     scaledLabelValue = parentScaler.scaleY(value: labelValue.toDouble());
   }
 
@@ -482,8 +423,7 @@ class Poly {
 }
 
 // todo 0 add tests; also make constant; also add validation for min before max
-// todo -11: parametrize with T
-// todo -11: replaced num with double
+// todo -2: replaced num with double,  parametrize with T instead so it works for both
 
 class Interval {
   Interval(this.min, this.max,